refactor ui trace graph overview
